requirements.txt
aiosqlite==0.19.0 ; python_version == "3.12"
alembic==1.16.2 ; python_version == "3.12"
annotated-types==0.7.0 ; python_version == "3.12"
anyio==4.9.0 ; python_version == "3.12"
argon2-cffi-bindings==21.2.0 ; python_version == "3.12"
argon2-cffi==23.1.0 ; python_version == "3.12"
bcrypt==4.2.0 ; python_version == "3.12"
certifi==2025.6.15 ; python_version == "3.12"
cffi==1.17.1 ; python_version == "3.12"
click==8.2.1 ; python_version == "3.12"
colorama==0.4.6 ; python_version == "3.12" and platform_system == "Windows"
cryptography==45.0.5 ; python_version == "3.12"
dnspython==2.7.0 ; python_version == "3.12"
email-validator==2.2.0 ; python_version == "3.12"
fastapi-csrf-protect==1.0.3 ; python_version == "3.12"
fastapi-users-db-sqlalchemy==7.0.0 ; python_version == "3.12"
fastapi-users==14.0.1 ; python_version == "3.12"
fastapi-users[sqlalchemy]==14.0.1 ; python_version == "3.12"
fastapi==0.115.14 ; python_version == "3.12"
greenlet==3.2.3 ; python_version == "3.12" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.16.0 ; python_version == "3.12"
httpcore==1.0.9 ; python_version == "3.12"
httpx==0.28.1 ; python_version == "3.12"
idna==3.10 ; python_version == "3.12"
itsdangerous==2.2.0 ; python_version == "3.12"
jinja2==3.1.6 ; python_version == "3.12"
makefun==1.16.0 ; python_version == "3.12"
mako==1.3.10 ; python_version == "3.12"
markupsafe==3.0.2 ; python_version == "3.12"
minio==7.2.15 ; python_version == "3.12"
nh3==0.2.21 ; python_version == "3.12"
prompt-toolkit==3.0.51 ; python_version == "3.12"
pwdlib[argon2,bcrypt]==0.2.1 ; python_version == "3.12"
pycparser==2.22 ; python_version == "3.12"
pycryptodome==3.23.0 ; python_version == "3.12"
pydantic-core==2.33.2 ; python_version == "3.12"
pydantic-settings==2.10.1 ; python_version == "3.12"
pydantic==2.11.7 ; python_version == "3.12"
pyjwt[crypto]==2.10.1 ; python_version == "3.12"
python-dotenv==1.1.1 ; python_version == "3.12"
python-multipart==0.0.20 ; python_version == "3.12"
sniffio==1.3.1 ; python_version == "3.12"
sqlalchemy==2.0.41 ; python_version == "3.12"
sqlalchemy[asyncio]==2.0.41 ; python_version == "3.12"
starlette==0.46.2 ; python_version == "3.12"
typing-extensions==4.14.1 ; python_version == "3.12"
typing-inspection==0.4.1 ; python_version == "3.12"
urllib3==2.5.0 ; python_version == "3.12"
uvicorn[standard]==0.35.0 ; python_version == "3.12"
wcwidth==0.2.13 ; python_version == "3.12"
poetry
loguru
passlib
aiofiles
pillow
python-multipart
piexif
jsonschema
python-docx
psutil
pymupdf
tuspy
aiofiles